Create Lab0.java
CIS 313, Intermediate Data Structures Winter 2017
CIS 313 Lab 0
Due: Jan 8th, 2016 at 11:59pm
This lab is an introduction into lab procedure and acts as a review of Java
Overview
For this assignment, you are to write a program which will read a series of pairs of integers X and Y and print pairs GCD(X, Y ) and LCM(X, Y ), where
• GCD(X, Y ): the greatest common divisor of X and Y , • LCM(X, Y ): the least common multiple of X and Y .
The purpose of this assignment is two-fold
• Practice using standard input and output. • Learn how to submit your work.
You should write your program in Java only.
Input Description
The input will be a text ﬁle, for example inSample.txt below will be provided. The ﬁrst line will contain an integer N, which is the number of lines to follow. Each of the N lines contains two integers X and Y , separated by a space.
5 4 2 8 9 143 247 1 1 10 10
Note: When using an editor, you may also manually type in input to the command window. However, you will be tested with a ﬁle similar to inSample.txt.
1

Output Description
For each of the pairs X Y output GCD(X, Y ) and LCM(X, Y ) on a line, separated by a single space. For example, using the sample input above, your program should output:
2 4 1 72 13 2717 1 1 10 10